+++ src/app.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { ListModule } from './modules/list/list.module';
-import { ItemModule } from './modules/item/item.module';
+import { ListModule } from './task-list/list.module';
+import { ItemModule } from './task-item/item.module';
 import { DatabaseModule } from './database.module';
 import { ConfigModule } from '@nestjs/config';
 import { validate } from './env.validate';
