+++ src/entities/base-entiy.abstract.ts
