+++ src/entities/list.entity.ts
@@ -1,6 +1,6 @@
 import { Column, Entity, OneToMany } from 'typeorm';
-import { Item } from '../item/item.entity';
-import { BaseEntity } from '../../common/abstracts/base-entiy.abstract';
+import { Item } from './item.entity';
+import { BaseEntity } from './base-entiy.abstract';
 
 @Entity()
 export class List extends BaseEntity {
